rename & test with c++11
