+++ src/AnalysisTreeQALinkDef.h
@@ -2,6 +2,7 @@
 
 #pragma link C++ namespace AnalysisTree::QA;
 #pragma link C++ class AnalysisTree::QA::Manager+;
+#pragma link C++ class AnalysisTree::QA::Axis+;
 #pragma link C++ class AnalysisTree::QA::EntryConfig+;
 
 #endif
