--- src/Axis.cxx
@@ -1,5 +1,0 @@
-//
-// Created by vklochkov on 3/24/20.
-//
-
-#include "Axis.h"
