+++ src/Manager.hpp
